fix: naming fixed, views and requests update to compilation events. Refactor: upload views and events not via manager but in service
