+++ main-service/src/main/java/ru/yandex/practicum/TestController.java
@@ -26,7 +26,6 @@ public class TestController {
                                               @RequestParam(name = "uris", defaultValue = "") ArrayList<String> uris,
                                               @RequestParam(name = "unique", defaultValue = "false") boolean unique) {
         log.info("TEST-GET \"/stats?start={}&end={}&uris={}&unique={}\"", start, end, uris, unique);
-        System.out.println("start: " + start + "\n" + "end: " + end);
         List<StatisticInfo> response = statisticClient.getAllByFilter(new StatisticFilterDto(start, end, uris, unique));
         log.debug("return: " + response.toString());
         return response;
