+++ main-service/src/main/java/ru/yandex/practicum/category/service/impl/CategoryServiceImpl.java
@@ -1,17 +1,20 @@
 package ru.yandex.practicum.category.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.category.Category;
 import ru.yandex.practicum.category.CategoryDto;
 import ru.yandex.practicum.category.CategoryMapper;
 import ru.yandex.practicum.category.CategoryRepository;
-import ru.yandex.practicum.category.service.CategoryServiceAdmin;
+import ru.yandex.practicum.category.service.CategoryService;
 import ru.yandex.practicum.exception.NotFoundException;
+
+import java.util.List;
 
 @RequiredArgsConstructor
 @Service
-public class CategoryServiceAdminImpl implements CategoryServiceAdmin {
+public class CategoryServiceImpl implements CategoryService {
     final CategoryRepository categoryRepository;
 
     @Override
@@ -35,5 +38,18 @@ public class CategoryServiceAdminImpl implements CategoryServiceAdmin {
         categoryRepository.findById(catId)
                 .orElseThrow(() -> new NotFoundException("Category with id=" + catId + " was not found"));
         categoryRepository.deleteById(catId);
+    }
+
+    @Override
+    public List<CategoryDto> getFiltered(Pageable pageable) {
+        List<Category> categoryList = categoryRepository.findAll(pageable).getContent();
+        return CategoryMapper.listToDtoList(categoryList);
+    }
+
+    @Override
+    public CategoryDto getById(long compId) {
+        Category category = categoryRepository.findById(compId)
+                .orElseThrow(() -> new NotFoundException("Category with id=" + compId + " was not found"));
+        return CategoryMapper.toDto(category);
     }
 }
