+++ main-service/src/main/java/ru/yandex/practicum/event/controller/EventPrivateController.java
@@ -10,7 +10,7 @@ import ru.yandex.practicum.event.model.dto.EventCreateDto;
 import ru.yandex.practicum.event.model.dto.EventFullInfoDto;
 import ru.yandex.practicum.event.model.dto.EventShortInfoDto;
 import ru.yandex.practicum.event.model.dto.EventUpdateDto;
-import ru.yandex.practicum.event.service.EventServicePrivate;
+import ru.yandex.practicum.event.service.EventService;
 import ru.yandex.practicum.eventRequest.model.EventRequestInfoDto;
 import ru.yandex.practicum.eventRequest.model.EventRequestStatusChanger;
 import ru.yandex.practicum.eventRequest.model.EventRequestStatusResult;
@@ -23,25 +23,25 @@ import java.util.List;
 @RequestMapping("/users/{userId}/events")
 @RequiredArgsConstructor
 @Validated
-public class EventControllerPrivate {
-    final EventServicePrivate eventServicePrivate;
+public class EventPrivateController {
+    final EventService eventService;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public EventFullInfoDto create(@PathVariable(name = "userId") long initiatorId,
                                    @RequestBody @Validated EventCreateDto eventCreateDto) {
         log.info("POST \"/user/{}/events\" Body={}", initiatorId, eventCreateDto);
-        EventFullInfoDto eventToReturn = eventServicePrivate.create(initiatorId, eventCreateDto);
+        EventFullInfoDto eventToReturn = eventService.create(initiatorId, eventCreateDto);
         log.debug("Event created=" + eventToReturn);
         return eventToReturn;
     }
 
     @PatchMapping("/{eventId}")
-    public EventFullInfoDto updateAsInitiator(@PathVariable(name = "userId") long initiatorId,
-                                              @PathVariable(name = "eventId") long eventId,
-                                              @RequestBody @Validated EventUpdateDto eventUpdateDto) {
+    public EventFullInfoDto update(@PathVariable(name = "userId") long initiatorId,
+                                   @PathVariable(name = "eventId") long eventId,
+                                   @RequestBody @Validated EventUpdateDto eventUpdateDto) {
         log.info("PATCH \"/user/{}/events/{}\" Body={}", initiatorId, eventId, eventUpdateDto);
-        EventFullInfoDto eventFullInfoDto = eventServicePrivate.updateAsInitiator(initiatorId, eventId, eventUpdateDto);
+        EventFullInfoDto eventFullInfoDto = eventService.update(initiatorId, eventId, eventUpdateDto);
         log.debug("Event updated=" + eventFullInfoDto);
         return eventFullInfoDto;
     }
@@ -52,7 +52,7 @@ public class EventControllerPrivate {
                                                             @RequestParam(name = "size", defaultValue = "10") int limit) {
         log.info("GET \"/user/{}/events&from={}&size={}\"", initiatorId, offset, limit);
         Pageable pageable = new OffsetPageable(offset, limit);
-        List<EventShortInfoDto> eventList = eventServicePrivate.getByInitiatorIdFiltered(initiatorId, pageable);
+        List<EventShortInfoDto> eventList = eventService.getByInitiatorIdFiltered(initiatorId, pageable);
         log.debug("EventList found=" + eventList);
         return eventList;
     }
@@ -61,7 +61,7 @@ public class EventControllerPrivate {
     public EventFullInfoDto getByIdAndInitiatorId(@PathVariable(name = "userId") long initiatorId,
                                                   @PathVariable(name = "eventId") long eventId) {
         log.info("GET \"/user/{}/events/{}", initiatorId, eventId);
-        EventFullInfoDto event = eventServicePrivate.getByIdAndInitiatorId(initiatorId, eventId);
+        EventFullInfoDto event = eventService.getByIdAndInitiatorId(initiatorId, eventId);
         log.debug("Event found=" + event);
         return event;
     }
@@ -70,18 +70,18 @@ public class EventControllerPrivate {
     public List<EventRequestInfoDto> getEventRequestsByInitiatorIdAndEventId(@PathVariable(name = "userId") long initiatorId,
                                                                              @PathVariable(name = "eventId") long eventId) {
         log.info("GET \"/user/{}/events/{}/requests", initiatorId, eventId);
-        List<EventRequestInfoDto> eventRequestList = eventServicePrivate.getEventRequestsByInitiatorIdAndEventId(initiatorId, eventId);
+        List<EventRequestInfoDto> eventRequestList = eventService.getEventRequestsByInitiatorIdAndEventId(initiatorId, eventId);
         log.debug("EventRequestList found=" + eventRequestList);
         return eventRequestList;
     }
 
 
     @PatchMapping("/{eventId}/requests")
-    public EventRequestStatusResult updateEventRequestsAsInitiator(@PathVariable(name = "userId") long initiatorId,
-                                                                   @PathVariable(name = "eventId") long eventId,
-                                                                   @RequestBody EventRequestStatusChanger eventRequestStatusChanger) {
+    public EventRequestStatusResult updateEventRequests(@PathVariable(name = "userId") long initiatorId,
+                                                        @PathVariable(name = "eventId") long eventId,
+                                                        @RequestBody EventRequestStatusChanger eventRequestStatusChanger) {
         log.info("PATCH \"/user/{}/events/{}/requests Body={}", initiatorId, eventId, eventRequestStatusChanger);
-        EventRequestStatusResult eventRequestList = eventServicePrivate.updateEventRequestsAsInitiator(initiatorId, eventId, eventRequestStatusChanger);
+        EventRequestStatusResult eventRequestList = eventService.updateEventRequests(initiatorId, eventId, eventRequestStatusChanger);
         log.debug("EventRequestList found=" + eventRequestList);
         return eventRequestList;
     }
