+++ main-service/src/main/java/ru/yandex/practicum/event/model/dto/EventCreateDto.java
@@ -32,8 +32,6 @@ public class EventCreateDto {
     String description;
 
     @NotNull(message = "Field: eventDate. Error: must not be null.")
-    //DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
-    //@DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     @InTwoHours(message = "Field: eventDate. Error: Must contain time at least 2 hours ahead of now")
     LocalDateTime eventDate;
