+++ main-service/src/main/java/ru/yandex/practicum/event/model/dto/EventFullInfoDto.java
@@ -30,7 +30,6 @@ public class EventFullInfoDto {
 
     String description;
 
-    //DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     LocalDateTime eventDate;
