+++ main-service/src/main/java/ru/yandex/practicum/event/model/dto/EventRequestAdminDto.java
@@ -28,7 +28,6 @@ public class EventRequestAdminDto {
     @Size(min = 20, max = 7000)
     String description;
 
-    //DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     @InTwoHours
