+++ main-service/src/main/java/ru/yandex/practicum/eventRequest/EventRequestRepository.java
@@ -8,16 +8,16 @@ import java.util.Collection;
 import java.util.List;
 
 public interface EventRequestRepository extends JpaRepository<EventRequest, Long> {
-    List<EventRequest> findAllByRequesterId(long requesterId);
+    List<EventRequest> findByRequesterId(long requesterId);
 
-    List<EventRequest> findAllByIdInAndEventId(List<Long> idList, long eventId);
+    List<EventRequest> findByIdInAndEventId(List<Long> idList, long eventId);
 
-    List<EventRequest> findAllByEventIdInAndStatus(Collection<Long> eventIdList, EventRequestStatus status);
+    List<EventRequest> findByEventIdInAndStatus(Collection<Long> eventIdList, EventRequestStatus status);
 
     long countByEventIdAndStatus(long eventId, EventRequestStatus status);
 
-    List<EventRequest> findAllByEventInitiatorIdAndEventId(long initiatorId, long eventId);
+    List<EventRequest> findByEventInitiatorIdAndEventId(long initiatorId, long eventId);
 
-    List<EventRequest> findAllByStatus(EventRequestStatus status);
+    List<EventRequest> findByStatus(EventRequestStatus status);
 
 }
