+++ main-service/src/main/java/ru/yandex/practicum/eventRequest/model/EventRequest.java
@@ -23,11 +23,11 @@ public class EventRequest {
     @Column(name = "event_request_id")
     Long id;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "event_id", nullable = false)
     Event event;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "user_id", nullable = false)
     User requester;
