+++ main-service/src/main/java/ru/yandex/practicum/user/UserRepository.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    List<User> findAllByIdIn(List<Long> users, Pageable pageable);
+    List<User> findByIdIn(List<Long> users, Pageable pageable);
 
-    List<User> findAllByIdIn(List<Long> users);
 }
